Simulate time series from the exponential Ornstein-Uhlenbeck stochastic volatility model.

Simulate time series from the exponential Ornstein-Uhlenbeck stochastic volatility model.

eou_sim(nobs, dt, X0, log_V0, alpha, log_gamma, mu, log_sigma, logit_rho, dBt)

Arguments

nobs

Length of time series.

dt

Interobservation time.

X0

Scalar or vector of nseries asset log prices at time t = 0.

log_V0

Scalar or vector of nseries volatilities at time t = 0, on the log standard deviation scale.

alpha

Scalar or vector of nseries growth rate parameters.

log_gamma

Scalar or vector of nseries log-volatility mean reversion parameters on the log scale.

mu

Scalar or vector of nseries log-volatility mean parameters.

log_sigma

Scalar or vector of nseries log-volatility diffusion parameters on the log scale.

logit_rho

Scalar or vector of nseries correlation parameters between asset and volatility innovations, on the logit scale.

dBt

An optional list with elements V and Z corresponding to matrices of size nobs x nseries of pre-specified Brownian innovations. If missing these consist of iid draws from an N(0, dt) distribution.

Value

A list containing matrices Xt and log_Vt of nobs x nseries of eOU observations, where each column corresponds to a process observed at times t = dt, 2dt, ..., nobs*dt.