Extract the Brownian increments from the SVC model.

Extract the Brownian increments from the SVC model.

svc_dB(Xt, log_VPt, log_Vt, dt, alpha, log_gamma, mu, log_sigma, logit_rho)

Arguments

Xt

Matrix of nobs x (nasset + 1) asset log prices, where the first column is that of the asset common-factor proxy.

log_VPt

Vector of nobs volatility proxy values on the log standard deviation scale. See 'Details'.

log_Vt

Optional vector of nobs x (nasset + 1) volatilities on the log standard deviation scale. See 'Details'.

dt

Interobservation time.

alpha

Optional vector of (nasset + 1) asset growth rate parameters. See 'Details'.

log_gamma

Optional vector of (nasset + 2) log-volatility mean reversion parameters on the log scale. The first two correspond to the volatility proxy and the common-factor asset's volatility, respectively. See 'Details'.

mu

Optional vector of (nasset + 2) log-volatility mean parameters. See 'Details'.

log_sigma

Optional vector of (nasset + 2) log-volatility diffusion parameters on the log scale. See 'Details'.

logit_rho

Optional vector of (nasset + 1) correlation parameters between asset and volatility innovations, on the logit scale. The first one is that of the common-factor asset proxy. See 'Details'.

Value

A list with elements:

V

An nobs x (nasset+2) matrix of the log-volatility innovations.

X

An nobs x (nasset+1) matrix of log-asset innovations.

Z

An nobs x (nasset+1) matrix of residual log-asset innovations, after account for the log-volatility innovations. That is, dB_Z = (dB_X - rho dB_V) / sqrt(1 - rho^2).